Migrated 'health' script to the HitScript
